Update status from endpoint

+++ src/main/kotlin/sh/zachwal/button/admin/contact/UpdateContactRequest.kt
@@ -1,6 +1,6 @@
 package sh.zachwal.button.admin.contact
 
 data class UpdateContactRequest(
-    val contactId: Long,
+    val contactId: Int,
     val active: Boolean
 )
